+++ noir-projects/noir-contracts/contracts/account/schnorr_initializerless_account_contract/Nargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "schnorr_initializerless_account_contract"
+type = "contract"
+compiler_version = ">=1.0.0"
+authors = [""]
+
+[dependencies]
+aztec = { path = "../../../../aztec-nr/aztec" }
+schnorr = { tag = "v0.4.0", git = "https://github.com/noir-lang/schnorr" }
+
