+++ noir-projects/noir-contracts/contracts/account/simulated_schnorr_initializerless_account_contract/Nargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "simulated_schnorr_initializerless_account_contract"
+type = "contract"
+compiler_version = ">=1.0.0"
+authors = [""]
+
+[dependencies]
+aztec = { path = "../../../../aztec-nr/aztec" }
