+++ yarn-project/accounts/package.json
@@ -8,16 +8,14 @@
     "./defaults": "./dest/defaults/index.js",
     "./ecdsa": "./dest/ecdsa/index.js",
     "./ecdsa/lazy": "./dest/ecdsa/lazy.js",
+    "./ecdsa/stub": "./dest/ecdsa/stub/index.js",
+    "./ecdsa/stub/lazy": "./dest/ecdsa/stub/lazy.js",
     "./schnorr": "./dest/schnorr/index.js",
     "./schnorr/lazy": "./dest/schnorr/lazy.js",
-    "./stub/schnorr": "./dest/stub/schnorr/index.js",
-    "./stub/schnorr/lazy": "./dest/stub/schnorr/lazy.js",
-    "./stub/ecdsa": "./dest/stub/ecdsa/index.js",
-    "./stub/ecdsa/lazy": "./dest/stub/ecdsa/lazy.js",
+    "./schnorr/stub": "./dest/schnorr/stub/index.js",
+    "./schnorr/stub/lazy": "./dest/schnorr/stub/lazy.js",
     "./testing": "./dest/testing/index.js",
     "./testing/lazy": "./dest/testing/lazy.js",
-    "./copy-cat": "./dest/copy_cat/index.js",
-    "./copy-cat/lazy": "./dest/copy_cat/lazy.js",
     "./utils": "./dest/utils/index.js"
   },
   "typedocOptions": {
