+++ yarn-project/accounts/src/defaults/stub_account_contract.ts
