+++ yarn-project/accounts/src/schnorr/account_contract.ts
@@ -4,6 +4,7 @@ import { Fr } from '@aztec/foundation/curves/bn254';
 import { GrumpkinScalar } from '@aztec/foundation/curves/grumpkin';
 import { AuthWitness } from '@aztec/stdlib/auth-witness';
 import { CompleteAddress } from '@aztec/stdlib/contract';
+import { deriveSigningKey } from '@aztec/stdlib/keys';
 
 import { DefaultAccountContract } from '../defaults/account_contract.js';
